# Improve performance of get_string