--- a/Cargo.toml
+++ b/Cargo.toml
@@ -33,6 +33,7 @@ java-locator = { version = "0.1", optional = true }
 jni-sys = "0.4"
 libloading = { version = "0.8", optional = true }
 log = "0.4.4"
+once_cell = "1.19.0"
 static_assertions = "1"
 thiserror = "2"