@benoitmeriaux 2024-10-31
@rib could you take a look and approve, it would be very nice to release this really appreciable speed bump on get_string :)