@rib 2025-03-04
Sorry for the delay with getting a chance to look at this.

I agree that it should be reasonable to be caching `java.lang.String` class references and even more generally think we should probably allow `jni-rs` to assume that the JVM is a global singleton that can't be re-initialized and look for other (separate from this PR) opportunities to take advantage of that assumption (e.g. `GlobalRef` perhaps shouldn't need to a copy of the `JavaVM` pointer and there are other `java.lang.` classes and method IDs we could cache), ref: https://github.com/jni-rs/jni-rs/issues/567

I've rebased the branch for this PR, and think we can merge this once it passes CI.