@olivergillespie 2025-03-04
Thanks! I actually have a new concern - https://mostlynerdless.de/blog/2023/07/17/jmethodids-in-profiling-a-tale-of-nightmares/ says that a jclass becomes invalid if the class is redefined. I haven't confirmed it applies to this case, but seems like it could be an issue. Redefining String is niche, of course, so I'm not sure if it's a practical concern.