@rib 2025-03-05
> Thanks! I actually have a new concern - https://mostlynerdless.de/blog/2023/07/17/jmethodids-in-profiling-a-tale-of-nightmares/ says that a jclass becomes invalid if the class is redefined. I haven't confirmed it applies to this case, but seems like it could be an issue. Redefining String is niche, of course, so I'm not sure if it's a practical concern.

Yeah, my initial concern with taking this direction where we start caching class references was that we'd have to start worrying about dynamic loading of classes but I don't _think_ we need to worry about `java.lang` classes which I believe can't normally be reloaded.

I had been pondering suggesting that we should centralise a more general, global class reference cache that would also support being explicitly cleared in case of dynamically loaded classes, but figured that for now we shouldn't have to worry about that kind of added complexity for this special case `String` class cache.

There is a notion of a "bootstrap class loader" that I believe should deal with loading classes like `java.lang.String` and then that would be sealed such that you can't arbitrarily monkey patch core classes like that in normal operation of a JVM.

There _may_ be some niche cases where JVMs could let you override those bootstrap classes but I'd guess even then that it wouldn't be possible to do dynamically.

I'm interested in following up with enabling more general caching for other `java.lang` classes (atm I feel like this `String` optimization is a semi-arbitrary special case) as well as method IDs, but to avoid some of the complexity from dynamic loading it could make sense that this would be strictly limited to `java.lang` classes - at least initially.